+++ R/trans_matrix.R
@@ -2,7 +2,8 @@
 #' @description
 #' The function transposes a species matrix, while preserving correct species and sample names. 
 #' The new column names must be stored as row names of the data frame. They may also be stored in the first column, when chosing the argument \code{row.names = F}. 
-#'
+#' Missing values (NA) will be transformed to 0.
+#' 
 #' @param matrix Community data, a data frame. 
 #' @param row.names A logical evaluation indicating whether the new column names are stored as row names of the data frame \code{TRUE} \emph{(default)} or in the first column \code{FALSE}.
 #' @param rmchar A logical evaluation indicating whether the first character of the original column names should be removed (default: \code{FALSE}).
@@ -27,8 +28,8 @@
 
 
 trans_matrix <- function(matrix, row.names = T, rmchar = FALSE) {
+  
   if(row.names == T) {
-    
     colnames <- names(matrix)
     if(rmchar == TRUE) colnames <- sub('.', '', colnames) 
     rownames <- row.names(matrix)                 # Row names as names
@@ -43,5 +44,9 @@ trans_matrix <- function(matrix, row.names = T, rmchar = FALSE) {
     matrix.trans <- data.frame(t(matrix)[-1,], row.names = colnames)
     names(matrix.trans) <- rownames
   }
+  
+  # Replace empty character values
+  matrix.trans[matrix.trans == ""] <- NA
+  
   return(matrix.trans)
 }